+++ aliases/testdata/wild/nested-wild/another/another/alias_test.txt
@@ -1,0 +1,1 @@
+EVEN_WILDER = 'someFlag'
